consistent semi-colons on imports; move common folder up one level
