+++ src/common/Geometry.js
@@ -1,4 +1,3 @@
-
 // A simple struct that we can use everywhere we need vertices. I don't see a
 // problem letting this bloat a little.
 //
@@ -19,4 +18,3 @@ export const degToRad = (deg) => {
 export const radToDeg = (rad) => {
   return rad * 180.0 / Math.PI;
 }
-
