+++ src/features/app/Documentation.css
@@ -1,4 +1,3 @@
-
 .documentation {
   padding: 10px;
   background-color: #222;
