+++ src/features/app/Header.css
@@ -1,4 +1,3 @@
-
 .App-header {
   background-color: #222;
   height: 190px;
