+++ src/features/machine/PolarSettings.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 import {
     Col,
     ControlLabel,
