+++ src/features/turtle/Slider.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 
 class Slider extends Component {
   constructor(props) {
